feat : adding the posibility to add dashboard vet
